Replaced Lookup table with linear function for feedforward

+++ sedani/speed_LUT_generation/speed_LUT_generation.ino
@@ -1,6 +1,6 @@
 //#define AUDIO_ENABLE
 
-const int set_PWM = 1500;
+const long set_PWM = 1500;
 
 #ifdef AUDIO_ENABLE
 #include "pitch.h"
